init wks version factory revision
to support 1.1.1 and 1.3.0
