+++ owslib/map/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: ISO-8859-15 -*-
+
+from __future__ import (absolute_import, division, print_function)
